+++ Hart-Engine/src/Utils/Random.hpp
@@ -17,7 +17,7 @@ namespace Hart {
 			int64_t getRandomInt64(int64_t begin, int64_t end);
 
 			float getRandomFloat(float begin, float end);
-			float getRandomDouble(double begin, double end);
+			double getRandomDouble(double begin, double end);
 		private:
 			std::unique_ptr<std::mt19937> m_MT19937;
 		};
